backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
PyJWT==2.8.0
pytz==2024.1
whitenoise==6.6.0
yfinance==0.2.36